Introducing Cursors in the Kernel API.
Cursors are now used to implement the DiskLayer operations.

+++ community/cypher/cypher-compiler-2.3/src/main/scala/org/neo4j/cypher/internal/compiler/v2_3/spi/RepeatableReadQueryContext.scala
@@ -68,12 +68,12 @@ final class RepeatableReadQueryContext(inner: QueryContext, locker: Locker) exte
   }
 
 
-  override def getPropertiesForNode(node: Long): Iterator[Long] = {
+  override def getPropertiesForNode(node: Long): Iterator[Int] = {
     lockNode(node)
     inner.getPropertiesForNode(node)
   }
 
-  override def getPropertiesForRelationship(relId: Long): Iterator[Long] = {
+  override def getPropertiesForRelationship(relId: Long): Iterator[Int] = {
     lockRelationship(relId)
     inner.getPropertiesForRelationship(relId)
   }
